ref: split messageID check
